+++ CHANGELOG.md
@@ -2,15 +2,40 @@
 
 (TODO correspond à l'identifiant d'un problème à traiter pour la prochaine version)
 
-## TODO - Blindage de la construction des arcs
+## TODO 0.2 - Ajout de fabriques pour les sommets et les arcs
+En ajoutant un constructeur Edge(source: Vertex, target: Vertex), on remarque que l'on a simplifié la création des arcs et des sommets.
 
-On remarque que l'on dispose d'un modèle qui ne protège pas contre les erreurs la création des sommets et des arcs. Il est tout à fait possible de créer par erreur un `Edge` avec une `source` ou une `target` nulle.
+Toutefois, les opérations de création demeurent complexes et il reste la possibilité d'oublier d'ajouter les éléments aux listes de vertices et edges du graphe.
 
-On procède comme suit :
+On va donc procéder comme suit :
 
-* [ ] Ajout d'un constructeur `Edge(source: Vertex, target: Vertex)`
-* [ ] Suppression du constructeur par défaut sur `Edge`
+ *  Ajout d'une fabrique createVertex(coordinate,id): Vertex dans Graph
+ *  Ajout d'une fabrique createEdge(source,target,id): Edge dans Graph
+ *  Masquage des constructeur Vertex() et Edge(source,target)
+ *  Suppression de Graph.setVertices(vertices) et Graph.setEdges(edges)
+ 
+Ainsi, on passera de :
 
+```java
+Vertex a = new Vertex();
+a.setId("a");
+a.setCoordinate(new Coordinate(0.0, 0.0));
+graph.getVertices().add(a);
 
+Vertex b = new Vertex();
+b.setId("b");
+b.setCoordinate(new Coordinate(1.0, 0.0));
+graph.getVertices().add(b);
 
+Edge ab = new Edge(a,b);
+ab.setId("ab");
+graph.getEdges().add(ab);
+```
+à
+
+```java
+Vertex a = graph.createVertex(new Coordinate(0.0, 0.0),"a");
+Vertex b = graph.createVertex(new Coordinate(1.0, 0.0),"b");
+Edge ab  = graph.createEdge(a,b,"ab");
+```